requirements.txt
pdfminer.six==20221105
pdfminer3k==1.3.4
